Pruned down to core streaming function, yaargs rebuilt, .ava upcoming
